% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsutil.R
\name{hh}
\alias{hh}
\title{A function to print the first n rows and columns of the input.}
\usage{
hh(x, n = 5, m = n)
}
\arguments{
\item{x}{A matrix or dataframe.}

\item{n}{Number of rows to display.}

\item{m}{Number of columns to display.}
}
\value{
Return the input and print its dimensions.
}
\description{
A function to print the first n rows and columns of the input.
}
\examples{
library(magrittr)
df <- data.frame(
 A = 1:10,
 B = letters[1:10],
 C = rnorm(10),
 D = rnorm(10)
)
hh(df, 2)
}
