% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsutil.R
\name{lw}
\alias{lw}
\title{A function to print the length of TRUE elements in a logical.}
\usage{
lw(x)
}
\arguments{
\item{x}{A vector or logicals.}
}
\value{
Returns the length of the elements which are TRUE in the input.
}
\description{
A function to print the length of TRUE elements in a logical.
}
\examples{
lw(c(TRUE, TRUE, FALSE, TRUE, FALSE))
}
