% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsutil.R
\name{pdim}
\alias{pdim}
\title{A function to return an object and printing its dimensions.}
\usage{
pdim(x)
}
\arguments{
\item{x}{A matrix or dataframe.}
}
\value{
Return the input and print its dimensions.
}
\description{
A function to return an object and printing its dimensions.
}
\examples{
library(magrittr)
df <- data.frame(
 A = 1:10,
 B = letters[1:10],
 C = rnorm(10),
 D = rnorm(10)
)
}
