% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsutil.R
\name{plength}
\alias{plength}
\title{A function to return an object and printing its length.}
\usage{
plength(x)
}
\arguments{
\item{x}{A vector or list.}
}
\value{
Returns the input and print its length.
}
\description{
A function to return an object and printing its length.
}
\examples{
library(magrittr)
v1 = 1:10 \%T>\% plength
}
