% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsutil.R
\name{pncol}
\alias{pncol}
\title{A function to return an object and printing its number of columns.}
\usage{
pncol(x)
}
\arguments{
\item{x}{A matrix or dataframe.}
}
\value{
Return the input and print its number of columns.
}
\description{
A function to return an object and printing its number of columns.
}
\examples{
library(magrittr)
df <- data.frame(
 A = 1:10,
 B = letters[1:10],
 C = rnorm(10),
 D = rnorm(10)
)
}
