% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsutil.R
\name{t2}
\alias{t2}
\title{A function to print a vector as a column.}
\usage{
t2(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
Returns the input transposed twice.
}
\description{
A function to print a vector as a column.
}
\examples{
library(magrittr)
t2(1:10)
}
